+++ drivers/interrupt_controller/intc_gic_common_priv.h
@@ -7,6 +7,45 @@
 #ifndef ZEPHYR_INCLUDE_DRIVERS_INTC_GIC_COMMON_PRIV_H_
 #define ZEPHYR_INCLUDE_DRIVERS_INTC_GIC_COMMON_PRIV_H_
 
+/*
+ * GIC Register Interface Base Addresses
+ */
+#define GIC_DIST_BASE	DT_REG_ADDR_BY_IDX(DT_INST(0, arm_gic), 0)
+#define GIC_CPU_BASE	DT_REG_ADDR_BY_IDX(DT_INST(0, arm_gic), 1)
+
+/*
+ * GIC Distributor Interface
+ */
+
+/*
+ * 0x000  Distributor Control Register
+ * v1		ICDDCR
+ * v2/v3	GICD_CTLR
+ */
+#define	GICD_CTLR		(GIC_DIST_BASE +   0x0)
+
+/*
+ * 0x004  Interrupt Controller Type Register
+ * v1		ICDICTR
+ * v2/v3	GICD_TYPER
+ */
+#define	GICD_TYPER		(GIC_DIST_BASE +   0x4)
+
+/*
+ * 0x008  Distributor Implementer Identification Register
+ * v1		ICDIIDR
+ * v2/v3	GICD_IIDR
+ */
+#define	GICD_IIDR		(GIC_DIST_BASE +   0x8)
+
+/*
+ * 0xF00  Software Generated Interrupt Register
+ * v1		ICDSGIR
+ * v2/v3	GICD_SGIR
+ */
+#define	GICD_SGIR		(GIC_DIST_BASE + 0xf00)
+
+
 /* Offsets from GICD base or GICR(n) SGI_base */
 #define GIC_DIST_IGROUPR		0x0080
 #define GIC_DIST_ISENABLER		0x0100
@@ -27,6 +66,7 @@
 #define ICENABLER(base, n)		(base + GIC_DIST_ICENABLER + (n) * 4)
 #define ISPENDR(base, n)		(base + GIC_DIST_ISPENDR + (n) * 4)
 #define ICPENDR(base, n)		(base + GIC_DIST_ICPENDR + (n) * 4)
+#define ICACTIVER(base, n)		(base + GIC_DIST_ICACTIVER + (n) * 4)
 #define IPRIORITYR(base, n)		(base + GIC_DIST_IPRIORITYR + n)
 #define ITARGETSR(base, n)		(base + GIC_DIST_ITARGETSR + (n) * 4)
 #define ICFGR(base, n)			(base + GIC_DIST_ICFGR + (n) * 4)
@@ -45,4 +85,49 @@
 				(GIC_GET_RDIST(GET_CPUID) + GICR_SGI_BASE_OFF) \
 				: GIC_DIST_BASE)
 #endif
+
+/* GICD_CTLR Interrupt group definitions */
+#define GICD_CTLR_ENABLE_G0		0
+#define GICD_CTLR_ENABLE_G1NS		1
+
+#if CONFIG_GIC_VER > 2
+#define GICD_CTLR_ENABLE_G1S		2
+#endif
+
+/* GICD_ICFGR */
+#define GICD_ICFGR_MASK			BIT_MASK(2)
+#define GICD_ICFGR_TYPE			BIT(1)
+/* GICD_TYPER.ITLinesNumber 0:4 */
+#define GICD_TYPER_ITLINESNUM_MASK	0x1f
+
+#define GIC_NUM_INTR_PER_REG		32
+
+#define GIC_NUM_CFG_PER_REG		16
+
+#define GIC_NUM_PRI_PER_REG		4
+
+#define GIC_NUM_TGT_PER_REG		4
+
+#define GIC_INT_32X_MASK		0xffffffffU
+
+/* GIC idle priority : value '0xff' will allow all interrupts */
+#define GIC_IDLE_PRIO			0xff
+
+/* Priority levels 0:255 */
+#define GIC_PRI_MASK			0xff
+
+/*
+ * '0xa0'is used to initialize each interrtupt default priority.
+ * This is an arbitrary value in current context.
+ * Any value '0x80' to '0xff' will work for both NS and S state.
+ * The values of individual interrupt and default has to be chosen
+ * carefully if PMR and BPR based nesting and preemption has to be done.
+ */
+#define GIC_INT_DEF_PRI_X4		0xa0a0a0a0
+
+/*
+ * GIC distributor configuration.
+ */
+void gic_dist_init(void);
+
 #endif /* ZEPHYR_INCLUDE_DRIVERS_INTC_GIC_COMMON_PRIV_H */
